requirements.txt
aiogram==3.7.0
APScheduler==3.10.4
asyncpg==0.29.0
fastapi==0.111.0
httpx==0.27.0
openai==1.35.3
pydantic-settings==2.3.4
python-dotenv==1.0.1
SQLAlchemy==2.0.31
uvicorn[standard]==0.30.1

